Initial small movement - extracting out React-only part of Javascript API
