+++ src/api/react/index.ts
@@ -1,0 +1,5 @@
+import { exhibit } from './exhibit'
+
+export { exhibit } from './exhibit'
+
+export default exhibit
