+++ src/ui-component-library/icon/index.exh.ts
@@ -1,4 +1,5 @@
-import exhibit, { simpleSelectModifier, simpleTextInputModifier } from '../../api'
+import { simpleSelectModifier, simpleTextInputModifier } from '../../api'
+import exhibit from '../../api/react'
 import Component, { DEFAULT_PROPS, NAME } from '.'
 import { GROUP_NAME } from '../common'
