+++ src/ui-component-library/number-input/index.exh.ts
@@ -1,4 +1,4 @@
-import exhibit from '../../api'
+import exhibit from '../../api/react'
 import Component, { NAME } from '.'
 import { GROUP_NAME } from '../common'
