+++ src/ui-component-library/run-button/index.exh.ts
@@ -1,4 +1,4 @@
-import exhibit from '../../api'
+import exhibit from '../../api/react'
 import Component, { DEFAULT_PROPS, NAME } from '.'
 import { GROUP_NAME } from '../common'
