+++ src/ui-component-library/test-result-count-summary/index.exh.ts
@@ -1,4 +1,5 @@
-import exhibit, { simpleNumberInputModifier } from '../../api'
+import { simpleNumberInputModifier } from '../../api'
+import exhibit from '../../api/react'
 import Component, { DEFAULT_PROPS, NAME } from '.'
 import { GROUP_NAME } from '../common'
