+++ version.txt
@@ -1,1 +1,1 @@
-2.2.28
+2.3.0-beta0
